+++ notebooks/dice.clj
@@ -15,7 +15,7 @@
                      (when side
                        [:div.mt-2 {:style {:font-size "6em"}} side])
                      [:button.bg-blue-500.hover:bg-blue-700.text-white.font-bold.py-2.px-4.rounded
-                      {:on-click (fn [e] (nextjournal.clerk.render/clerk-eval '(roll!)))} "Roll 🎲!"]])}
+                      {:on-click (fn [e] (nextjournal.clerk.render/clerk-eval {:recompute? true} '(roll!)))} "Roll 🎲!"]])}
 @dice
 
 ;; Our roll! function `resets!` our `dice` with a random side and prints and says the result. Finally it updates the notebook.