+++ src/nextjournal/clerk/sci_env.cljs
@@ -1,4 +1,5 @@
 (ns nextjournal.clerk.sci-env
+  (:refer-clojure :exclude [time])
   (:require-macros [nextjournal.clerk.render.macros :refer [sci-copy-nss]])
   (:require ["@codemirror/language" :as codemirror-language]
             ["@codemirror/state" :as codemirror-state]
@@ -8,11 +9,12 @@
             ["framer-motion" :as framer-motion]
             ["react" :as react]
             [applied-science.js-interop :as j]
-            [cljs.reader]
             [cljs.math]
+            [cljs.reader]
             [clojure.string :as str]
             [edamame.core :as edamame]
             [goog.object]
+            [nextjournal.clerk.cherry-env :as cherry-env]
             [nextjournal.clerk.parser]
             [nextjournal.clerk.render :as render]
             [nextjournal.clerk.render.code]
@@ -24,6 +26,7 @@
             [nextjournal.clojure-mode.commands]
             [nextjournal.clojure-mode.extensions.eval-region]
             [nextjournal.clojure-mode.keymap]
+            [reagent.dom.server :as dom-server]
             [sci.configs.applied-science.js-interop :as sci.configs.js-interop]
             [sci.configs.reagent.reagent :as sci.configs.reagent]
             [sci.core :as sci]
@@ -63,7 +66,7 @@
 (defn ->viewer-eval-with-error [form]
   (try (*eval* form)
        (catch js/Error e
-         (js/console.error "error in viewer-eval" e)
+         (js/console.error "error in viewer-eval" e form)
          (ex-info (str "error in viewer-eval: " (.-message e)) {:form form} e))))
 
 (defonce !edamame-opts
@@ -75,8 +78,11 @@
          :read-cond :allow
          :readers
          (fn [tag]
-           (or (get {'viewer-fn   ->viewer-fn-with-error
-                     'viewer-eval ->viewer-eval-with-error} tag)
+           (or (get @cljs.reader/*tag-table* tag)
+               (get {'viewer-fn ->viewer-fn-with-error
+                     'viewer-fn/cherry cherry-env/->viewer-fn-with-error
+                     'viewer-eval ->viewer-eval-with-error
+                     'viewer-eval/cherry cherry-env/->viewer-eval-with-error} tag)
                (fn [value]
                  (viewer/with-viewer `viewer/tagged-value-viewer
                    {:tag tag
@@ -90,10 +96,12 @@
 (defn ^:export read-string [s]
   (edamame/parse-string s @!edamame-opts))
 
+(defn read-string-without-tag-table [s]
+  (binding [cljs.reader/*tag-table* (atom {})]
+    (edamame/parse-string s @!edamame-opts)))
 
 (def ^{:doc "Stub implementation to be replaced during static site generation. Clerk is only serving one page currently."}
-  doc-url
-  (sci/new-var 'doc-url (fn [x] (str "#" x))))
+  doc-url (sci/new-var 'doc-url viewer/doc-url))
 
 (def viewer-namespace
   (merge (sci/copy-ns nextjournal.clerk.viewer (sci/create-ns 'nextjournal.clerk.viewer))
@@ -101,6 +109,7 @@
           'doc-url doc-url
           'url-for render/url-for
           'read-string read-string
+          'read-string-without-tag-table read-string-without-tag-table
           'clerk-eval render/clerk-eval
           'consume-view-context view-context/consume
           'inspect-presented render/inspect-presented
@@ -119,10 +128,16 @@
 
 (def core-ns (sci/create-ns 'clojure.core nil))
 
+(defn ^:sci/macro time [_ _ expr]
+  `(let [start# (system-time)
+         ret# ~expr]
+     (prn (cljs.core/str "Elapsed time: "
+                         (.toFixed (- (system-time) start#) 6)
+                         " msecs"))
+     ret#))
+
 (def initial-sci-opts
-  {:async? true
-   :disable-arity-checks true
-   :classes {'js (j/assoc! goog/global "import" shadow.esm/dynamic-import)
+  {:classes {'js (j/assoc! goog/global "import" shadow.esm/dynamic-import)
              'framer-motion framer-motion
              :allow :all}
    :js-libs {"@codemirror/language" codemirror-language
@@ -135,7 +150,9 @@
    :ns-aliases '{clojure.math cljs.math}
    :namespaces (merge {'nextjournal.clerk.viewer viewer-namespace
                        'clojure.core {'read-string read-string
-                                      'implements? (sci/copy-var implements?* core-ns)}}
+                                      'implements? (sci/copy-var implements?* core-ns)
+                                      'time (sci/copy-var time core-ns)
+                                      'system-time (sci/copy-var system-time core-ns)}}
                       (sci-copy-nss
                        'cljs.math
                        'nextjournal.clerk.parser
@@ -151,16 +168,17 @@
                       sci.configs.js-interop/namespaces
                       sci.configs.reagent/namespaces)})
 
+(defn ^:export onmessage [ws-msg]
+  (render/dispatch (read-string (.-data ws-msg))))
+
 (defn ^:export eval-form [f]
   (sci/eval-form (sci.ctx-store/get-ctx) f))
 
-(defn ^:export set-state [state]
-  (render/set-state! state))
+(def ^:export init render/init)
 
-(def ^:export mount render/mount)
-
-(defn ^:export onmessage [ws-msg]
-  (render/dispatch (read-string (.-data ws-msg))))
+(defn ^:export ssr [state-str]
+  (init (read-string state-str))
+  (dom-server/render-to-string [render/root]))
 
 (defn reconnect-timeout [failed-connection-attempts]
   (get [0 0 100 500 5000] failed-connection-attempts 10000))