+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Changes can be:
 * 🌟⭐️💫 features
+* 🚨 possibly breaking
 * 🐞🐜 friendly or nasty bugs
 * 🛠 dev improvements
 
@@ -11,10 +12,45 @@ Changes can be:
 
     To see what's going on while waiting for a long-running computation, Clerk will now show an execution status bar on the top. For named cells (defining a var) it will show the name of the var, for anonymous expressions, a preview of the form.
 
+* 🔗 Interactive Links, Index and Homepage
+
+    Links can now be followed in interactive mode and the index can be viewed. Previously this could only be seen after a `build!`. Add support evaluating a given doc by entering it in the browser's address bar.
+
+    Use these features to build a new welcome page that gives more useful information, including links to potential notebooks in the project.
+
 * 🍕 `clerk/fragment` for splicing a seq of values into the document as if it were produced by results of individual cells. Useful when programmatically generating content.
 
+* 🚨 Change `nextjournal.clerk.render/clerk-eval` to not recompute the currently shown document when using the 1-arity version. Added a second arity that takes an opts map with a `:recompute?` key.
+
+* 🍒 Add support for cherry as an alternative to sci to evaluate `:render-fn`s. You can change it per form (using form metadata or viewer opts) or doc-wide (using ns metadata) with `{:nextjournal.clerk/render-evaluator :cherry}`.
+
+* ⭐️ Adds support for customization of viewer options
+
+  Support both globally (via ns metadata or a settings marker) or locally (via form metadata or the viewer options map).
+
+  Supported options are:
+   * `:nextjournal.clerk/auto-expand-results?`
+   * `:nextjournal.clerk/budget`
+   * `:nextjournal.clerk/css-class`
+   * `:nextjournal.clerk/visibility`
+   * `:nextjournal.clerk/width`
+   * `:nextjournal.clerk/render-evaluator`
+
 * 🔌 Make websocket reconnect automatically on close to avoid having to reload the page
+
 * 💫 Cache expressions that return `nil` in memory
+
+* 💫 Support non-evaluated clojure code listings in markdown documents by specifying `{:nextjournal.clerk/code-listing true}` after the language ([#482](https://github.com/nextjournal/clerk/issues/482)).
+
+* 🏳️‍🌈 Syntax highlighting for code listings in all [languages supported by codemirror](https://github.com/codemirror/language-data) ([#500](https://github.com/nextjournal/clerk/issues/500)).
+
+* 🐜 Turn off analyzer pass for validation of `:type` tags, fixes [#488](https://github.com/nextjournal/clerk/issues/488) @craig-latacora
+
+* 🐜 Strip `:type` metadata from forms before printing them to hash, fixes [#489](https://github.com/nextjournal/clerk/issues/489) @craig-latacora
+
+* 🐜 Ensure custom `print-method` supporting unreadable symbols preserves metadata
+
+* 🐞 Preserve `*ns*` during `build!`, fixes [#506](https://github.com/nextjournal/clerk/issues/506)
 
 ## 0.13.842 (2023-03-07)